Move all public functions to top-level Toolshed module
