+++ lib/toolshed/tree.ex
@@ -1,29 +1,9 @@
 defmodule Toolshed.Tree do
-  @moduledoc """
-  This module provides the `tree` command
-  """
+  @moduledoc false
 
-  @doc """
-  Print out directories and files in tree form.
-  """
-  @spec tree(Path.t()) :: :"do not show this result in output"
-  def tree(path \\ ".") do
-    IO.puts(path)
+  def do_tree(_prefix, _dir, []), do: :ok
 
-    case file_info(path, path) do
-      {:directory, _} ->
-        do_tree("", path, files(path))
-
-      _ ->
-        :ok
-    end
-
-    IEx.dont_display_result()
-  end
-
-  defp do_tree(_prefix, _dir, []), do: :ok
-
-  defp do_tree(prefix, dir, [{:directory, filename} | rest]) do
+  def do_tree(prefix, dir, [{:directory, filename} | rest]) do
     puts_tree_branch(prefix, filename, rest)
 
     path = Path.join(dir, filename)
@@ -31,9 +11,19 @@ defmodule Toolshed.Tree do
     do_tree(prefix, dir, rest)
   end
 
-  defp do_tree(prefix, dir, [{_type, filename} | rest]) do
+  def do_tree(prefix, dir, [{_type, filename} | rest]) do
     puts_tree_branch(prefix, filename, rest)
     do_tree(prefix, dir, rest)
+  end
+
+  def files(dir) do
+    File.ls!(dir)
+    |> Enum.map(&file_info(Path.join(dir, &1), &1))
+  end
+
+  def file_info(path, name) do
+    stat = File.lstat!(path)
+    {stat.type, name}
   end
 
   defp puts_tree_branch(prefix, filename, rest) do
@@ -45,14 +35,4 @@ defmodule Toolshed.Tree do
 
   defp tree_trunk([]), do: "    "
   defp tree_trunk(_), do: "│   "
-
-  defp files(dir) do
-    File.ls!(dir)
-    |> Enum.map(&file_info(Path.join(dir, &1), &1))
-  end
-
-  defp file_info(path, name) do
-    stat = File.lstat!(path)
-    {stat.type, name}
-  end
 end
