+++ test/toolshed/lsof_test.exs
@@ -2,6 +2,6 @@ defmodule Toolshed.LsofTest do
   use ExUnit.Case
 
   test "lsof/0 returns correct value" do
-    assert Toolshed.Lsof.lsof() == :ok
+    assert Toolshed.lsof() == :ok
   end
 end
