+++ test/toolshed/misc_test.exs
@@ -8,7 +8,7 @@ defmodule Toolshed.MiscTest do
     path = tmp_file(context.tmp_dir, "text_data")
     content = {:some_interesting_atom, ["some", "list"]}
 
-    assert Toolshed.Misc.save_value(content, path) == :ok
+    assert Toolshed.save_value(content, path) == :ok
     assert File.read!(path) == "{:some_interesting_atom, [\"some\", \"list\"]}"
   end
 
@@ -16,8 +16,8 @@ defmodule Toolshed.MiscTest do
     path = tmp_file(context.tmp_dir, "binary_data")
     content = {:some_interesting_atom, ["some", "list"]}
 
-    assert Toolshed.Misc.save_term!(content, path) == content
-    assert Toolshed.Misc.load_term!(path) == content
+    assert Toolshed.save_term!(content, path) == content
+    assert Toolshed.load_term!(path) == content
   end
 
   defp tmp_file(tmp_dir, filename) do
