+++ test/toolshed/weather_test.exs
@@ -1,11 +1,10 @@
 defmodule Toolshed.WeatherTest do
   use ExUnit.Case
-  import Toolshed.Weather
 
   test "weather/0 ensures :inets and :ssl are started" do
     Application.stop(:inets)
     Application.stop(:ssl)
 
-    assert weather() == :"do not show this result in output"
+    assert Toolshed.weather() == :"do not show this result in output"
   end
 end
